Increase pcap offsets to 64 bits to prevent overflow as per bug 0045.
darcs-hash:20070919004622-f1522-0ec4613faa1d91eb4f3b15acc17ad528c8d5b2d6.gz

+++ src/include/pcap.h
@@ -47,7 +47,7 @@ struct pcap_pkthdr {
   uint32_t len;	/* length this packet (off wire) */
   char *data;
 
-  uint32_t offset;
+  uint64_t offset;
   uint32_t id;
   // This holds the dissected tree
   Root root;
@@ -68,7 +68,7 @@ CLASS(PcapFileHeader, Packet)
      char *le_format;
 
      // The offset into the pcap file of this packet
-     unsigned long long int pcap_offset;
+     uint64_t pcap_offset;
 END_CLASS
 
 CLASS(PcapPacketHeader, Packet)
